requirements.txt

python-dotenv
google-generative-ai
openai